requirements.txt
tqdm~=4.64.0
transformers~=4.23.1
pyvi~=0.1.1
numpy~=1.21.0
scikit-learn~=1.0.2
setuptools~=58.0.4
datasets~=2.6.1
redis~=4.4.0
torch~=1.13.0
python-dotenv~=0.21.0
pydantic~=1.9.2
levenshtein~=0.20.8
uvicorn~=0.20.0
fastapi~=0.88.0
wandb~=0.13.7